fake freq values

+++ TremAn3/ViewModels/FreqCounterViewModel.cs
@@ -30,7 +30,7 @@ namespace TremAn3.ViewModels
         public List<DataPoint> DataPoints { get; set; } = new List<DataPoint>();// { new DataPoint(1, 10), new DataPoint(2, 11), new DataPoint(3, 9) };
 
         //convertor from tuple to datapoinnts
-        internal void UpdatePlotWithNewVals(IEnumerable<(int xx, int yy)> newVals)
+        internal void UpdatePlotWithNewVals(IEnumerable<(double xx, double yy)> newVals)
         {
             DataPoints.Clear();
             newVals.ToList().ForEach(c => DataPoints.Add(new DataPoint(c.xx, c.yy)));
@@ -77,5 +77,7 @@ namespace TremAn3.ViewModels
             string TimeString = result.ToString("mm':'ss");
             return TimeString;
         }
+
+
     }
 }
